requirements.txt
# Core RAG
langchain>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
rank-bm25>=0.2.2
sentence-transformers>=2.7.0

# Server & UI
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# MCP server (will be added later)

# Optional: LLM generation
langchain-text-splitters>=0.2.0
